backend/requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
pydantic==1.10.17
python-dotenv==1.0.0
requests==2.30.0
# Fixando versões compatíveis para evitar erros de incompatibilidade binária
numpy==1.22.4
scipy==1.8.1
spacy==3.5.3
mysql-connector-python==8.0.33
python-multipart==0.0.6
jwt==1.3.1
pyjwt==2.7.0
sentence-transformers==2.2.2
